more cleaning the code
